refactor: break out IndexingError to own file
